Add files via upload
this code creates random keys between 0 and 105 for the English letters plus space. it does that by creating x random numbers where x is provided in the project description. it then uses jmodl to encrypt the plain text by replacing each character in the plain text with j mod l where j is the index of the letter in plain text and l is the key list length for that letter.